requirements.txt
asgiref==3.2.4
bootstrap4==0.1.0
Django==3.0.4
django-crispy-forms==1.10.0
flake8==3.8.4
gunicorn==20.0.4
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.2.0
